Remove unnecessary JSON.stringify in GCP handler
When I inspected responses from the GCP handler in Postman, I noticed
that sometimes it's pretty printed and sometimes not. I realized that
some of the messages correctly set the response header (to JSON). It
seems GCP automatically sets this header when the body is an object so
I've removed the stringification.