Unify cloud providers gateways error handling (#1292)
* Remove unnecessary error handling

These checks are already performed by AWS and having them in Airnode
is just confusing and are dead code.

* Synchronize and improve error handling of AWS and GCP gateways

* Remove unnecessary airnode-node code

* Add changeset

* Minor refactor

* Do not perform unnecessary validation in AWS handlers

* Add unit tests, fix bugs

* Remove unnecessary JSON.stringify in GCP handler

When I inspected responses from the GCP handler in Postman, I noticed
that sometimes it's pretty printed and sometimes not. I realized that
some of the messages correctly set the response header (to JSON). It
seems GCP automatically sets this header when the body is an object so
I've removed the stringification.