Remove network host when running docker container
This option is only available for Linux and it breaks the API server
which is used for local gateways.

"""
The host networking driver only works on Linux hosts, and is not
supported on Docker Desktop for Mac, Docker Desktop for Windows, or
Docker EE for Windows Server.
"""
(From https://docs.docker.com/network/host/)