Remove unnecessary async/await (#1439)
* Remove unnecessary await, enforce it via eslint

* Remove unnecessary "async", enforce it via eslint

* Add changeset

* Use "mockRejectedValue" instead of "mockImplementation"

This is needed, because we are mocking failures of a third party
(ethers) async functions. Previously we used async mocks, but that was
confusing since the functions never awaited anything (just threw the
error). When the async was removed the functions started throwing
synchronous errors. These errors were not handled in ethers (because
they can never happen) and we got strange jest errors instead.